Move legacy and refactored legacy to separate crates
