+++ .github/workflows/build_nix.yml
@@ -11,12 +11,6 @@ on:
 
 jobs:
   nix:
-    strategy:
-      matrix:
-        flags: ["--cfg legacy_builder_refactored", ""]
-    env:
-      RUSTFLAGS: "${{ matrix.flags }}"
-      RUSTDOCFLAGS: "${{ matrix.flags }}"
     runs-on: ubuntu-latest
     timeout-minutes: 90
     steps:
