+++ Cargo.toml
@@ -3,7 +3,7 @@ resolver = "2"
 members = ["crates/*"]
 
 [workspace.package]
-version = "0.1.57"
+version = "0.1.58"
 edition = "2021"
 
 [workspace.dependencies]
