+++ crates/legacy/src/builder_state.rs
