+++ crates/legacy/src/testing/finalization_test.rs
