+++ crates/legacy_refactored/src/block_store.rs
